Soem refacoring
